MNIST original
The MNIST dataset has the array of images that are stored in form of an array. The idea is to train on the dataset using Decision Tree Classifier algorithm and predict the number which is present on the given location of the array. and to resize them and display it.